% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggnomogram.R, R/ggnomogramV3.R
\name{which_near}
\alias{which_near}
\title{plot nomogram from logistic regression modeling}
\usage{
which_near(x, comp_vector = c(0.1, 0.5, 1, 2, 5, 10, 20, 25, 50, 100, 250,
  1000, 10000, 1e+05, 1e+06, 1e+07, 1e+08))

which_near(x, comp_vector = c(0.1, 0.5, 1, 2, 5, 10, 20, 25, 50, 100, 250,
  1000, 10000, 1e+05, 1e+06, 1e+07, 1e+08))
}
\arguments{
\item{model_info}{list of model info see example.}

\item{model_info}{list of model info see example.}
}
\description{
plot nomogram from logistic regression modeling

plot nomogram for estimation of sum of
}
